remove needless ComboBox abstraction
It was only used by the ModelSelectField.
